+++ src/main/java/com/shop/admin/product/dto/ProductAdminUpdateRequest.java
@@ -2,11 +2,9 @@ package com.shop.admin.product.dto;
 
 import jakarta.validation.constraints.Size;
 
-import java.math.BigDecimal;
-
 public record ProductAdminUpdateRequest(
         @Size(max = 200) String name,
-        BigDecimal price,
+        int price,
         @Size(max = 4000) String description,
         @Size(max = 500) String mainImageUrl,
         Long categoryId
