+++ src/main/java/com/shop/admin/sales/dto/SalesDailyItem.java
@@ -1,11 +1,10 @@
 package com.shop.admin.sales.dto;
 
-import java.math.BigDecimal;
 
 public record SalesDailyItem(
         String productName,
-        BigDecimal price,
+        int price,
         int quantity,
-        BigDecimal amount
+        long amount
 ) {
 }
